+++ packages/api/.nycrc
@@ -4,5 +4,5 @@
   "include": ["src/**/*.ts", "src/**/*.js"],
   "reporter": ["text", "html", "cobertura", "lcov"],
   "check-coverage": true,
-  "lines": 90
+  "lines": 75
 }
