--- packages/api/tests/ui.test.ts
@@ -1,75 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import "mocha";
-import { Result } from "neverthrow";
-import { Colors, FxError } from "../src";
-import {
-  InputTextConfig,
-  InputTextResult,
-  IProgressHandler,
-  MultiSelectConfig,
-  MultiSelectResult,
-  SelectFileConfig,
-  SelectFileResult,
-  SelectFilesConfig,
-  SelectFilesResult,
-  SelectFolderConfig,
-  SelectFolderResult,
-  SingleSelectConfig,
-  SingleSelectResult,
-  UserInteraction,
-} from "../src/qm/ui";
-
-async function sleep(ms: number) {
-  await new Promise((resolve) => setTimeout(resolve, ms));
-}
-class MockUserInteraction implements UserInteraction {
-  selectOption(config: SingleSelectConfig): Promise<Result<SingleSelectResult, FxError>> {
-    throw new Error("Method not implemented.");
-  }
-  selectOptions(config: MultiSelectConfig): Promise<Result<MultiSelectResult, FxError>> {
-    throw new Error("Method not implemented.");
-  }
-  inputText(config: InputTextConfig): Promise<Result<InputTextResult, FxError>> {
-    throw new Error("Method not implemented.");
-  }
-  selectFile(config: SelectFileConfig): Promise<Result<SelectFileResult, FxError>> {
-    throw new Error("Method not implemented.");
-  }
-  selectFiles(config: SelectFilesConfig): Promise<Result<SelectFilesResult, FxError>> {
-    throw new Error("Method not implemented.");
-  }
-  selectFolder(config: SelectFolderConfig): Promise<Result<SelectFolderResult, FxError>> {
-    throw new Error("Method not implemented.");
-  }
-
-  openUrl(link: string): Promise<Result<boolean, FxError>> {
-    throw new Error("Method not implemented.");
-  }
-  async showMessage(
-    level: "info" | "warn" | "error",
-    message: string,
-    modal: boolean,
-    ...items: string[]
-  ): Promise<Result<string | undefined, FxError>>;
-
-  async showMessage(
-    level: "info" | "warn" | "error",
-    message: Array<{ content: string; color: Colors }>,
-    modal: boolean,
-    ...items: string[]
-  ): Promise<Result<string | undefined, FxError>>;
-
-  async showMessage(
-    level: "info" | "warn" | "error",
-    message: string | Array<{ content: string; color: Colors }>,
-    modal: boolean,
-    ...items: string[]
-  ): Promise<Result<string | undefined, FxError>> {
-    throw new Error("Method not implemented.");
-  }
-  createProgressBar(title: string, totalSteps: number): IProgressHandler {
-    throw new Error("Method not implemented.");
-  }
-}
